LayoutTests/platform/chromium/accessibility/range-for-line-textarea-expected.txt


This tests that a popup button will return the text under the element for its title.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS axElement.rangeForLine(0) is '{0, 14}'
PASS axElement.rangeForLine(1) is '{14, 14}'
PASS axElement.rangeForLine(2) is '{28, 16}'
PASS axElement.rangeForLine(3) is '{44, 14}'
PASS axElement.rangeForLine(4) is '{58, 14}'
PASS axElement.rangeForLine(6) is '{84, 14}'
PASS axElement.rangeForLine(7) is '{98, 16}'
PASS axElement.rangeForLine(8) is '{114, 14}'
PASS axElement.rangeForLine(9) is '{128, 12}'
PASS axElement.rangeForLine(10) is '{140, 14}'
PASS axElement.rangeForLine(100) is '{0, 0}'
PASS axElement.rangeForLine(-1) is '{0, 0}'
PASS successfullyParsed is true

TEST COMPLETE
